Make gem update prompt quieter
The Ruby LSP will prompts users to update the gem if it isn't
already up-to-date with an intrusive pop-up. This has been distracting
and confusing users. Instead, we can implement this functionality via an
unobtrusive addition to the status bar. When the user clicks the button
in the status bar, it will update the gem for them automatically and
then disappear.

Add progress bar to indicate LSP upgrade progress