Use `spawn` for better progress reporting during `bundle install`
The `exec` call buffers up the contents of stdout and stderr, which it
then returns as a single string when the process exits. This means that
we don't get any progress reporting until the process exits.

By using `spawn` instead, we can get progress reporting as the process
writes to its stdout and stderr streams. We hook the `data` event on
the `stdout` and `stderr` streams of the `ChildProcess` object, and
we output that information to our output panel.

However, it would be good to prefix the output with some kind of string
when dumping it to the output panel, so that we can distinguish it from
other messages. In order to do that, we pipe the `stdout` and `stderr`
streams through `readline` so that we get a stream of lines, which
we can then prefix with a string.

Finally, when the `ChildProcess` sends us the `close` event, we clean up
our `readline` objects and fulfill or reject the promise returned from
the method.